temporary fix of minor datatype issue

+++ gpaw/mpi/__init__.py
@@ -533,7 +533,9 @@ class _Communicator:
             return self.comm.translate_ranks(other.comm, ranks) # argh!
         otherranks = self.comm.translate_ranks(other.get_c_object(), ranks)
         assert all(-1 <= rank for rank in otherranks)
-        return otherranks
+        # XXX the C code returns some kind of NPY_ integer type which is
+        # not necessarily the default, so we convert to the default one:
+        return otherranks.astype(int)
         
     def get_members(self):
         """Return the subset of processes which are members of this MPI group
